resolved the requests from the PR and removed unnecessary files
